feat: Added transparency and order of rendering accordingly to camera distance

+++ Engine/src/Engine/Core/Constants.h
@@ -21,8 +21,8 @@ namespace MyEngine
 	const glm::vec4 UP_VECTOR = glm::vec4(0, 1, 0, 1);
 	const glm::vec4 FORWARD_VECTOR = glm::vec4(0, 0, -1, 1);
 
-	const int    FRAME_RATE = 60;  // Frames per second
-	const double FRAME_DURATION = 1.0 / FRAME_RATE;
+	const int   FRAME_RATE = 60;  // Frames per second
+	const float FRAME_DURATION = 1.0f / (float)FRAME_RATE;
 
 	const glm::vec4 RED = glm::vec4(1.0f, 0.0f, 0.0f, 1.0f);
 	const glm::vec4 GREEN = glm::vec4(0.0f, 1.0f, 0.0f, 1.0f);
